YLW-3176 refactor(CalculatorTests): Combines tests files correctly
